docker/wifi_audit/requirements.txt
scapy==2.4.4
Django==3.2
django_celery_results==2.0.1
djangorestframework==3.12.4
django-crontab==0.7.1
celery==5.0.5
requests==2.25.1
python3-nmap==1.5.0
billiard==3.6.4.0
psycopg2==2.8.6
